requirement.txt
alembic==1.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
gevent==23.9.1
greenlet==3.0.3
idna==3.6
install==1.3.5
Mako==1.3.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
websocket==0.2.1
websockets==12.0
zope.event==5.0
zope.interface==6.1